backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
torch>=2.2.0
torchvision>=0.17.0
pillow==10.1.0
numpy==1.26.1
python-dotenv==1.0.0